Add files via upload
KPIs are critical for any firm to track, monitor, and analyze crucial aspects in order to improve marketing performance over time. One of the most cost-effective methods to expand your business is to track the correct B2B SaaS metrics and make changes. So, now that you know how important it is to track KPIs for effective B2B campaigns, the next step is to arrange the data, which is usually done using business intelligence tools or dashboards. As a result, it's critical to develop and execute smarter campaigns while measuring vital data in order to propel your company to new heights.